academy_backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
mongoengine==0.27.0
python-dotenv==1.1.1
pymongo==3.12.3
django-cors-headers==4.4.0
gunicorn==21.2.0
dj-database-url==2.2.0
whitenoise[brotli]==6.6.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
dnspython==2.8.0
werkzeug==3.1.1
djangorestframework-simplejwt
razorpay>=2.0.0
google-auth
reportlab
